fix companion device pairing
